src/requirements.txt
Django==5.1.4
django-environ==0.11.2
django-ninja==1.3.0
djangorestframework==3.15.2
Faker==33.1.0
locust==2.32.5
pytest==8.3.4
pytest-django==4.9.0
psycopg2-binary>=2.9.6